add predator eating logic

+++ src/main/java/com/rush/service/FeedingService.java
@@ -1,10 +1,14 @@
 package com.rush.service;
 
+import com.rush.config.AnimalRegistry;
 import com.rush.domain.map.Cell;
 import com.rush.domain.orgaism.Organism;
 import com.rush.domain.orgaism.animal.Animal;
+import com.rush.domain.orgaism.animal.herbivore.Herbivore;
+import com.rush.domain.orgaism.animal.predator.Predator;
 
 import java.util.List;
+import java.util.concurrent.ThreadLocalRandom;
 
 public class FeedingService {
     private final CellService cellService;
@@ -20,6 +24,10 @@ public class FeedingService {
                 continue;
             }
 
+            if (animal instanceof Predator predator && !isVictimCaught(predator, (Herbivore) organism)) {
+                continue;
+            }
+
             animal.eat(organism);
             cellService.removeOrganism(cell, organism);
 
@@ -29,4 +37,9 @@ public class FeedingService {
         }
     }
 
+    private static boolean isVictimCaught(Predator predator, Herbivore victim) {
+        int probability = AnimalRegistry.getProbabilityToCatch(predator.getClass(), victim.getClass());
+
+        return ThreadLocalRandom.current().nextInt(100) < probability;
+    }
 }
